+++ src/Order/OrderService.php
@@ -255,4 +255,29 @@ final class OrderService extends AbstractService
             throw $e;
         }
     }
+
+    /**
+     * @param int $orderId
+     * @return OrderAdjustment[]
+     * @throws AuthenticationRequired
+     * @throws NotFound
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     */
+    public function getAdjustments(int $orderId): array
+    {
+        $this->client->mustBeAuthenticated();
+
+        try {
+            return array_map(function (array $data): OrderAdjustment {
+                return new OrderAdjustment($data);
+            }, $this->client->get("orders/{$orderId}/adjustments"));
+        } catch (ClientException $e) {
+            if ($e->getResponse()->getStatusCode() === 404) {
+                throw new NotFound("Order #{$orderId} not found", $e);
+            }
+
+            throw $e;
+        }
+    }
 }